update_TN027 | UI | RUD_Gallery

+++ centerlearn-common/src/main/java/com/nhatanh/centerlearn/common/response/GalleryResponse.java
@@ -16,5 +16,5 @@ public class GalleryResponse {
     private final String description;
     private final LocalDateTime createdAt;
     private final LocalDateTime updatedAt;
-    private final double fileSize;
+    private final long fileSize;
 }
